Add new rule SrcDuplicate to language.json
This new rule is for deduplication. The syntax is [["preffered_form", "form2", "form3"]].
Working Example: "SrcDuplicate":[["á", "ā"]] in Slovak
- This after recompile of goruut will phonemize ā as if it were á (the preferred form).
- Can also specify multiple rules.

Regarding study, study must be run with this SrcDuplicate already set to the correct
substrings. This will then make grammar (Map) so that only preffered form is in use.
All other forms will be replaced to the preferred form.

Same in inference, other forms will be replaced to the preferred form.

See issue #11